requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.5.1
aom=3.6.0=h313beb8_0
appnope=0.1.4=py311hca03da5_0
asttokens=3.0.0=py311hca03da5_0
blas=1.0=openblas
bottleneck=1.4.2=py311hb9f6ed7_0
brotli-python=1.0.9=py311h313beb8_9
brotlicffi=1.0.9.2=py311h313beb8_1
bzip2=1.0.8=h80987f9_6
c-ares=1.19.1=h80987f9_0
ca-certificates=2025.6.15=hbd8a1cb_0
cairo=1.16.0=hd886ae0_6
certifi=2025.6.15=py311hca03da5_0
cffi=1.17.1=py311h3eb5a62_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.1=py311hca03da5_0
contourpy=1.3.1=py311h48ca7d4_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h9131b1a_1
dav1d=1.2.1=h80987f9_0
debugpy=1.8.11=py311h313beb8_0
decorator=5.1.1=pyhd3eb1b0_0
deprecated=1.2.13=py311hca03da5_0
eigen=3.4.0=h48ca7d4_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.7.1=h313beb8_0
ffmpeg=6.1.1=h7693a39_3
filelock=3.17.0=py311hca03da5_0
fontconfig=2.14.1=h6402c1e_3
fonttools=4.55.3=py311h80987f9_0
freetype=2.13.3=h47d26ad_0
fribidi=1.0.10=h1a28f6b_0
fsspec=2025.5.1=py311hd096484_0
gdk-pixbuf=2.42.10=h969f7ef_2
gettext=0.21.0=hbdbcc25_2
giflib=5.2.2=h80987f9_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
glib=2.84.2=hc880cf1_0
glib-tools=2.84.2=hc880cf1_0
gmp=6.3.0=h313beb8_0
gmpy2=2.2.1=py311h5c1b81f_0
graphite2=1.3.14=hc377ac9_1
gst-plugins-base=1.24.12=hc962c03_1
gstreamer=1.24.12=h4a1bc71_1
gstreamer-orc=0.4.41=hdbbdb2f_0
harfbuzz=10.2.0=he637ebf_1
hdf5=1.14.5=hd77251f_2
icu=73.1=h313beb8_0
idna=3.7=py311hca03da5_0
importlib-metadata=8.5.0=py311hca03da5_0
ipykernel=6.29.5=py311hca03da5_1
ipython=9.1.0=py311hca03da5_0
ipython_pygments_lexers=1.1.1=py311hca03da5_0
jedi=0.19.2=py311hca03da5_0
jinja2=3.1.6=py311hca03da5_0
jpeg=9e=h80987f9_3
jupyter_client=8.6.3=py311hca03da5_0
jupyter_core=5.8.1=py311hca03da5_0
kiwisolver=1.4.8=py311h313beb8_0
krb5=1.20.1=hf3e1bf2_1
lame=3.100=h1a28f6b_0
lcms2=2.16=he26ebf3_1
leptonica=1.82.0=hdedf53e_3
lerc=4.0.0=h313beb8_0
libabseil=20250127.0=cxx17_h313beb8_0
libarchive=3.7.7=h8f13d7a_0
libcurl=8.12.1=hde089ae_0
libcxx=17.0.6=he5c5206_4
libdeflate=1.22=h80987f9_0
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libglib=2.84.2=hdc2269c_0
libiconv=1.16=h80987f9_3
libnghttp2=1.57.0=h62f6fdd_0
libogg=1.3.5=h1a28f6b_1
libopenblas=0.3.29=hea593b9_0
libopus=1.3.1=h80987f9_1
libpng=1.6.39=h80987f9_0
libpq=17.4=h02f6b3c_0
libprotobuf=5.29.3=h14f15fd_1
librsvg=2.56.3=h38920c8_2
libsodium=1.0.18=h1a28f6b_0
libssh2=1.11.1=h3e2b118_0
libtheora=1.1.1=h1a28f6b_3
libtiff=4.7.0=h91aec0a_0
libtorch=2.5.1=gpu_mps_h7b1bc93_204
libuv=1.48.0=h80987f9_0
libvorbis=1.3.7=h1a28f6b_0
libvpx=1.13.1=h313beb8_0
libwebp=1.3.2=hf40cdb4_1
libwebp-base=1.3.2=h80987f9_1
libxml2=2.13.8=h0b34f26_0
llvm-openmp=17.0.6=h2f17746_0
lz4-c=1.9.4=h313beb8_1
markupsafe=3.0.2=py311h80987f9_0
matplotlib=3.10.0=py311hca03da5_0
matplotlib-base=3.10.0=py311h7ef442a_0
matplotlib-inline=0.1.6=py311hca03da5_0
mpc=1.3.1=h80987f9_0
mpfr=4.2.1=h80987f9_0
mpmath=1.3.0=py311hca03da5_0
mysql=8.4.0=h065ec36_2
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=py311hca03da5_0
networkx=3.4.2=py311hca03da5_0
nomkl=3.0=0
numexpr=2.10.1=py311h5d9532f_0
numpy=2.3.1=py311hc3f7a0e_0
numpy-base=2.3.1=py311he1c501e_0
opencv=4.10.0=py311h122a3da_6
openh264=2.1.1=h98b2900_0
openjpeg=2.5.2=hba36e21_1
openldap=2.6.4=he7ef289_0
openssl=3.5.1=h81ee809_0
opentelemetry-api=1.30.0=py311hca03da5_0
packaging=24.2=py311hca03da5_0
pandas=2.2.3=py311hcf29cfe_0
pango=1.50.7=hf2328af_2
parso=0.8.4=py311hca03da5_0
patsy=1.0.1=pyhd8ed1ab_1
pcre2=10.42=hb066dcc_1
pexpect=4.9.0=py311hca03da5_0
pillow=11.1.0=py311h41ba818_1
pip=25.1=pyhc872135_2
pixman=0.40.0=h1a28f6b_0
platformdirs=4.3.7=py311hca03da5_0
prompt-toolkit=3.0.43=py311hca03da5_0
prompt_toolkit=3.0.43=hd3eb1b0_0
psutil=5.9.0=py311h80987f9_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-cpuinfo=9.0.0=pyhd8ed1ab_1
pycparser=2.21=pyhd3eb1b0_0
pygments=2.19.1=py311hca03da5_0
pyparsing=3.2.0=py311hca03da5_0
pysocks=1.7.1=py311hca03da5_0
python=3.11.13=h19e8193_0
python-dateutil=2.9.0post0=py311hca03da5_2
python-tzdata=2025.2=pyhd8ed1ab_0
pytorch=2.5.1=gpu_mps_py311hbd1d365_204
pytz=2025.2=pyhd8ed1ab_0
pyyaml=6.0.2=pyhe1237c8_2
pyzmq=26.2.0=py311h313beb8_0
qtbase=6.7.3=hb2d0045_0
readline=8.2=h1a28f6b_0
requests=2.32.4=py311hca03da5_0
scipy=1.15.3=py311hb81b3df_0
seaborn=0.13.2=hd8ed1ab_3
seaborn-base=0.13.2=pyhd8ed1ab_3
setuptools=72.1.0=py311hca03da5_0
six=1.17.0=py311hca03da5_0
sleef=3.5.1=h80987f9_2
smmap=5.0.2=pypi_0
sqlite=3.45.3=h80987f9_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.14.4=py311h80987f9_0
sympy=1.13.3=py311hca03da5_1
tesseract=5.2.0=h313beb8_2
tk=8.6.14=h6ba3021_1
torchvision=0.20.1=mps_py311h620d032_102
tornado=6.5.1=py311h80987f9_0
tqdm=4.67.1=pyhd8ed1ab_1
traitlets=5.14.3=py311hca03da5_0
typing_extensions=4.12.2=py311hca03da5_0
tzdata=2025b=h04d1e81_0
ultralytics=8.3.163=pyh2a12c56_0
unicodedata2=15.1.0=py311h80987f9_1
urllib3=2.5.0=py311hca03da5_0
wcwidth=0.2.13=py311hca03da5_0
wheel=0.45.1=py311hca03da5_0
wrapt=1.17.0=py311h80987f9_0
xz=5.6.4=h80987f9_1
yaml=0.2.5=h3422bc3_2
zeromq=4.3.5=h313beb8_0
zipp=3.21.0=py311hca03da5_0
zlib=1.2.13=h18a0788_1
zstd=1.5.6=hfb09047_0
